Add Vipps Api Client
