+++ src/Sjop/Services/Vipps/IVippsApiClient.cs
@@ -1,0 +1,11 @@
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Sjop.Services.Vipps
+{
+    public interface IVippsApiClient
+    {
+        Task<string> GetAccessToken(VippsSettings vippsSettings);
+
+    }
+}
