+++ src/Sjop/appsettings.json
@@ -22,6 +22,7 @@
     "WebhookSecret": "whsec_asdf"
   },
   "Vipps": {
+    "BaseUrl": "https://apitest.vipps.no",
     "MerchantSerialNumber": "1234",
     "ClientId": "asdf-asdf-asdf-asdf-asdfasdfasdf",
     "ClientSecret": "QWER1234",
